+++ src/main/java/com/daily/themoti/smokearea/domain/SmokeArea.java
@@ -1,4 +1,4 @@
-package com.daily.themoti.smokearea;
+package com.daily.themoti.smokearea.domain;
 
 import lombok.Builder;
 import lombok.Getter;
