+++ src/main/java/com/daily/themoti/smokearea/dto/SmokeAreaResponseDto.java
@@ -1,6 +1,6 @@
 package com.daily.themoti.smokearea.dto;
 
-import com.daily.themoti.smokearea.SmokeArea;
+import com.daily.themoti.smokearea.domain.SmokeArea;
 import lombok.Getter;
 
 @Getter
