+++ src/main/java/com/daily/themoti/smokearea/repository/SmokeAreaRepository.java
@@ -1,9 +1,7 @@
 package com.daily.themoti.smokearea.repository;
 
-import com.daily.themoti.smokearea.SmokeArea;
+import com.daily.themoti.smokearea.domain.SmokeArea;
 import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.List;
 
 public interface SmokeAreaRepository extends JpaRepository<SmokeArea, Long> {
